enum 타입으로 관리하기

+++ src/test/java/com/example/study/repository/ItemRepositoryTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.transaction.Transactional;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class ItemRepositoryTest extends StudyApplicationTests {
@@ -21,7 +22,7 @@ public class ItemRepositoryTest extends StudyApplicationTests {
         item.setName("삼성 노트북");
         item.setTitle("삼성 노트북 A100");
         item.setContent("2019년형 노트북 입니다.");
-        item.setPrice(900000);
+        item.setPrice(BigDecimal.valueOf(900000));
         item.setBrandName("삼성");
         item.setRegisteredAt(LocalDateTime.now());
         item.setCreatedAt(LocalDateTime.now());
